Move native bundle tests to E2E
