+++ scripts/testing-v2/test-map-execution.mjs
@@ -14,6 +14,7 @@ export const APPROVED_E2E_VITEST_PATHS = Object.freeze([
 	"tests2/core/team-manager.test.ts",
 	"tests2/integration/base-path-cli-entrypoint.test.ts",
 	"tests2/integration/benchmark-bobbit-journeys.test.ts",
+	"tests2/integration/pack-native-assets-bundle.test.ts",
 	"tests2/integration/team-spawn-multi-repo-real-git.test.ts",
 ]);
